Add Imenu support
Finding out the keywords or test cases using imenu is slightly more convoluted
than just a regular expression. Any string beginning from the first column of
the row is a keyword if it is in either the Keywords or Test Cases section.

https://robotframework.org/robotframework/latest/RobotFrameworkUserGuide.html#test-case-syntax

Add also ERT tests for the discovery of Imenu entries.

This should address #14